+++ lca/src/main/java/com/dreampany/lca/data/source/remote/CoinRemoteDataSource.java
@@ -158,12 +158,12 @@ public class CoinRemoteDataSource implements CoinDataSource {
     }
 
     @Override
-    public int getCount() {
+    public long getCount() {
         return 0;
     }
 
     @Override
-    public Maybe<Integer> getCountRx() {
+    public Maybe<Long> getCountRx() {
         return null;
     }
