+++ lca/src/main/java/com/dreampany/lca/data/source/remote/ExchangeRemoteDataSource.java
@@ -46,12 +46,12 @@ public class ExchangeRemoteDataSource implements ExchangeDataSource {
     }
 
     @Override
-    public int getCount() {
+    public long getCount() {
         return 0;
     }
 
     @Override
-    public Maybe<Integer> getCountRx() {
+    public Maybe<Long> getCountRx() {
         return null;
     }
